ready to merge in configure_storage.go
